Added edition rights for authors of observations and confirmations

+++ krokus-app/krokus-api/Controllers/ConfirmationsController.cs
@@ -4,6 +4,7 @@ using krokus_api.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace krokus_api.Controllers
 {
@@ -12,17 +13,13 @@ namespace krokus_api.Controllers
     public class ConfirmationsController : ControllerBase
     {
         private readonly IConfirmationService _confirmationService;
+        private readonly IAuthorizationService _authorizationService;
 
-        public ConfirmationsController(IConfirmationService confirmationService)
+        public ConfirmationsController(IConfirmationService confirmationService, IAuthorizationService authorizationService)
         {
             _confirmationService = confirmationService;
+            _authorizationService = authorizationService;
         }
-
-        /*
-        public async Task<ActionResult<IEnumerable<ConfirmationDto>>> GetAllConfirmations()
-        {
-            return Ok(await _confirmationService.FindAllConfirmations());
-        }*/
 
         [HttpGet]
         [AllowAnonymous]
@@ -47,42 +44,63 @@ namespace krokus_api.Controllers
         [Authorize(Policy = Policies.HasUserRights)]
         public async Task<ActionResult<ConfirmationDto>> PostConfirmation(ConfirmationDto confDto)
         {
+            if (confDto.UserId == null || confDto.UserId != User.FindFirstValue(ClaimTypes.NameIdentifier))
+            {
+                return BadRequest("UserId must be the same as the id of currently logged-in user.");
+            }
             var createdConf = await _confirmationService.CreateConfirmation(confDto);
             return Ok(createdConf);
         }
 
         [HttpPut("{id}")]
-        [Authorize(Policy = Policies.HasModeratorRights)]
         public async Task<IActionResult> PutConfirmation(long id, ConfirmationDto confDto)
         {
             if (id != confDto.Id)
             {
                 return BadRequest();
             }
-            bool result = await _confirmationService.UpdateConfirmation(confDto);
-
-            if (result)
+            var currentConf = await _confirmationService.FindById(id);
+            if (currentConf == null)
             {
+                return NotFound();
+            }
+            var authResult = await _authorizationService.AuthorizeAsync(User, currentConf, Policies.IsAuthorOrHasModeratorRights);
+            if (authResult.Succeeded)
+            {
+                bool result = await _confirmationService.UpdateConfirmation(confDto);
                 return NoContent();
+            }
+            else if (User.Identity?.IsAuthenticated ?? false)
+            {
+                return Forbid();
             }
             else
             {
-                return NotFound();
+                return Unauthorized();
             }
         }
 
         [HttpDelete("{id}")]
-        [Authorize(Policy = Policies.HasModeratorRights)]
         public async Task<IActionResult> DeleteConfirmation(long id)
         {
-            var result = await _confirmationService.DeleteConfirmation(id);
-            if (result)
+            var currentConf = await _confirmationService.FindById(id);
+            if (currentConf == null)
             {
+                return NotFound();
+            }
+            var authResult = await _authorizationService.AuthorizeAsync(User, currentConf, Policies.IsAuthorOrHasModeratorRights);
+            if (authResult.Succeeded)
+            {
+                var result = await _confirmationService.DeleteConfirmation(id);
                 return NoContent();
+            }
+            else if (User.Identity?.IsAuthenticated ?? false)
+            {
+                return Forbid();
             }
             else
             {
-                return NotFound();
+                return Unauthorized();
             }
         }
     }
